+++ components/gorgias/sources/common/event-types.mjs
@@ -1,0 +1,9 @@
+export default {
+  USER_CREATED: "user-created",
+  ACCOUNT_CREATED: "account-created",
+  CUSTOMER_CREATED: "customer-created",
+  TICKET_CREATED: "ticket-created",
+  TICKET_UPDATED: "ticket-updated",
+  TICKET_MESSAGE_CREATED: "ticket-message-created",
+  TICKET_MESSAGE_FAILED: "ticket-message-failed",
+};
